refactor: moved llm-related services to own package
