+++ src/main/java/se/ltu/eduo/llm/LlmService.java
@@ -1,4 +1,4 @@
-package se.ltu.eduo.service;
+package se.ltu.eduo.llm;
 
 public interface LlmService {
     String generateStudyQuestions(String prompt);
