+++ src/main/java/se/ltu/eduo/llm/PromptService.java
@@ -1,4 +1,4 @@
-package se.ltu.eduo.service;
+package se.ltu.eduo.llm;
 
 import org.springframework.stereotype.Service;
